% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{fit_data}
\alias{fit_data}
\title{Fitting right censored survival data to distribution}
\usage{
fit_data(data, dist, time = "Time", censor = "Censor", by = "")
}
\arguments{
\item{data}{A dataframe containing a time column and a censor column.}

\item{dist}{A string name for a distribution that has a corresponding density function and a distribution function.
Examples include "norm", "lnorm", "exp", "weibull", "logis", etc.}

\item{time}{The string name of the time column of the dataframe. Defaults to "Time".}

\item{censor}{The string name of the censor column of the dataframe. Defaults to "Censor". 
The censor column must be a numeric indicator variable where complete times correspond 
to a value of 1 and incomplete times correspond to 0.}

\item{by}{The string name of a grouping variable. if specified, the function returns a list. 
Variable can contain logical, string, character, or numeric data.}
}
\description{
Fits right censored data to a distribution using maximum liklihood estimates.
}
\examples{
library(survival) 
data("rats")
fit_data(rats, "lnorm", time = "time", censor = "status")

fit_data(rats, "weibull", time = "time", censor = "status", by = "sex")

}
\seealso{
\code{\link[fitdistrplus]{fitdistcens}}
}
