% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plotting density function overlayed on top of a histogram of data}
\usage{
plot_density(data, dist, time = "Time", censor = "Censor")
}
\arguments{
\item{data}{a dataframe containing a time column and a censor column.}

\item{dist}{a string name for a distribution that has a corresponding density function and distribution function.
Examples include "norm", "lnorm", "exp", "weibull", "logis", "llogis", "gompertz", etc.}

\item{time}{the string name of the time column of the dataframe. Defaults to "Time".}

\item{censor}{The string name of the censor column of the dataframe. Defaults to "Censor". 
The censor column must be a numeric indicator variable where complete times correspond 
to a value of 1 and incomplete times correspond to 0.}
}
\description{
Creates histogram of right censored data with the density function of a fitted parametric distribution overlayed.
}
\examples{
library(survival) 
data("kidney")
plot_density(kidney, "logis", time = "time", censor = "status")
plot_density(kidney, "weibull", time = "time", censor = "status")
}
