% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfuncs.R
\docType{package}
\name{survfuncs}
\alias{survfuncs}
\alias{survfuncs-package}
\title{survfuncs: Fitting right censored data to parametric distributions.}
\description{
Executes parametric survival analysis techniques similar to those in Minitab. Fits right censored data to a
given parametric distribution, produces summary statistics of the fitted distribution, and plots parametric survival, 
hazard, and cumulative hazard plots. Produces Anderson-Darling test statistic and probability plots to assess goodness
of fit of right censored data to a distribution.
}
